src/grech/pyproject.toml
[tool.poetry]
name = "grech"
version = "0.1.0"
description = "The yaml file can be freely written, and regular expressions can be tested using shell. In addition, when the test fails, an arbitrary message can be output."
authors = ["Yusuke Ohmori <you@example.com>"]
packages = [
    { include = "grech" },
    { include = "grech/**/*.py" },
]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
fire = "^0.4.0"
rich = "^12.5.1"
black = "^22.6.0"
pydantic = "^1.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=1.00", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
grech = "grech.main:main"